+++ Lab1/PictureDraw/PictureDraw/Tetragons.cs
@@ -13,19 +13,19 @@ namespace PictureDraw
     [Serializable]
     public class Tetragons : Shapes
     {
-        private int X1 { get; set; }
-        private int Y1 { get; set; }
-        private int X2 { get; set; }
-        private int Y2 { get; set; }
-        private int X3 { get; set; }
-        private int Y3 { get; set; }
-        private int X4 { get; set; }
-        private int Y4 { get; set; }
+        private float X1 { get; set; }
+        private float Y1 { get; set; }
+        private float X2 { get; set; }
+        private float Y2 { get; set; }
+        private float X3 { get; set; }
+        private float Y3 { get; set; }
+        private float X4 { get; set; }
+        private float Y4 { get; set; }
 
         public Tetragons() { }
 
         public Tetragons(string Name,
-            int startX, int startY, int finishX, int finishY): base(
+            float startX, float startY, float finishX, float finishY): base(
                 Name)
         {
             //finish not initialize 
@@ -76,7 +76,7 @@ namespace PictureDraw
     class TetragonCreator : ICreator
     {
         public  Shapes FactoryMethod(string Name,
-            int startX, int startY, int finishX, int finishY)
+            float startX, float startY, float finishX, float finishY)
         {
             return new Tetragons(Name, startX, startY, finishX, finishY);
         }
